+++ src/actions.ts
@@ -1,62 +1,23 @@
-/* ========================================================================== */
-/*                                   ACTIONS                                  */
-/* ========================================================================== */
+import { commands, TextEditorEdit, TextLine, window } from 'vscode';
 
-import { commands, TextEditor, TextEditorEdit, TextLine } from 'vscode';
-
-import { renderSubHeader, renderMainHeader, renderLine } from './renders';
+import {
+  renderHeader,
+  renderLine,
+  wrapWithMargins,
+  wrapWithLinebreaker
+} from './renders';
 import { checkEmptyLine } from './errors';
-import { NEW_LINE_SYM } from './constants';
-
-///
-
-export type Action = (editor: TextEditor, line: TextLine, lang: string) => void;
-
-interface IMargins {
-  top: boolean;
-  bottom: boolean;
-}
-
-/* --------------------------------- Margins -------------------------------- */
-
-const computeMargins = (line: TextLine, editor: TextEditor): IMargins => {
-  const isEmptyLine = (lineNum: number) =>
-    editor.document.lineAt(lineNum).isEmptyOrWhitespace;
-
-  const lastLineNum = editor.document.lineCount - 1;
-  const prevLineNum = line.lineNumber - 1;
-  const nextLineNum = line.lineNumber + 1;
-  const margins = {
-    top: false,
-    bottom: false
-  };
-
-  margins.top = prevLineNum >= 0 && !isEmptyLine(prevLineNum);
-  margins.bottom = nextLineNum <= lastLineNum && !isEmptyLine(nextLineNum);
-
-  return margins;
-};
-
-///
-
-const wrapMargins = (rendered: string, margins: IMargins): string => {
-  const before: string = margins.top ? NEW_LINE_SYM : '';
-  const after: string = margins.bottom ? NEW_LINE_SYM : '';
-
-  return before + rendered + after;
-};
+import { Action } from './types';
 
 /* --------------------------------- Inserts -------------------------------- */
 
-export const insertSubHeader: Action = (editor, line, lang) => {
+export const insertMainHeader: Action = (line, lang) => {
   checkEmptyLine(line);
 
-  const currentText = line.text.trim();
-  const rendered = renderSubHeader(currentText, lang);
-  const margins = computeMargins(line, editor);
-  const content = wrapMargins(rendered, margins);
+  const rendered = renderHeader('mainHeader', line.text, lang);
+  const content = wrapWithMargins(rendered, line);
 
-  editor
+  window.activeTextEditor
     .edit((textEditorEdit: TextEditorEdit) => {
       textEditorEdit.replace(line.range, content);
     })
@@ -67,15 +28,13 @@ export const insertSubHeader: Action = (editor, line, lang) => {
 
 ///
 
-export const insertMainHeader: Action = (editor, line, lang) => {
+export const insertSubHeader: Action = (line, lang) => {
   checkEmptyLine(line);
 
-  const currentText = line.text.trim();
-  const rendered = renderMainHeader(currentText, lang);
-  const margins = computeMargins(line, editor);
-  const content = wrapMargins(rendered, margins);
+  const rendered = renderHeader('subheader', line.text, lang);
+  const content = wrapWithMargins(rendered, line);
 
-  editor
+  window.activeTextEditor
     .edit((textEditorEdit: TextEditorEdit) => {
       textEditorEdit.replace(line.range, content);
     })
@@ -86,11 +45,11 @@ export const insertMainHeader: Action = (editor, line, lang) => {
 
 ///
 
-export const insertSolidLine: Action = (editor, line, lang) => {
+export const insertSolidLine: Action = (line, lang) => {
   const rendered = renderLine(lang);
-  const content = rendered + NEW_LINE_SYM;
+  const content = wrapWithLinebreaker(rendered);
 
-  editor
+  window.activeTextEditor
     .edit((textEditorEdit: TextEditorEdit) => {
       textEditorEdit.insert(line.range.start, content);
     })